+++ Plugins/Tum4ik.JustClipboardManager.PluginsDevKit/Services/IPluginSettingsService.cs
@@ -1,0 +1,9 @@
+using System.ComponentModel;
+using System.Globalization;
+
+namespace Tum4ik.JustClipboardManager.PluginDevKit.Services;
+public interface IPluginSettingsService
+{
+  CultureInfo Language { get; set; }
+  string Theme { get; set; }
+}