+++ Tum4ik.JustClipboardManager/Services/Translation/TranslationService.cs
@@ -2,7 +2,6 @@ using System.Collections.Immutable;
 using System.Resources;
 using Prism.Events;
 using Tum4ik.JustClipboardManager.PluginDevKit.Events;
-using Tum4ik.JustClipboardManager.PluginDevKit.Services;
 using Tum4ik.JustClipboardManager.Resources.Icons;
 
 namespace Tum4ik.JustClipboardManager.Services.Translation;