+++ Tum4ik.JustClipboardManager/Views/Main/Settings/SettingsPasteWindowView.xaml.cs
@@ -1,0 +1,11 @@
+namespace Tum4ik.JustClipboardManager.Views.Main.Settings;
+/// <summary>
+/// Interaction logic for SettingsPasteWindowView.xaml
+/// </summary>
+public partial class SettingsPasteWindowView
+{
+  public SettingsPasteWindowView()
+  {
+    InitializeComponent();
+  }
+}