+++ ReleaseNotes.md
@@ -3,6 +3,7 @@
 * Keep user settings from the previous version of the application after upgrade.
 * Automatic restart on application crash.
 * "Remove clips older than..." setting.
+* Paste window settings.
 
 ### Fixes
 * Fixed problem when the application breaks the clipboard viewer chain on exit.